Remove init and replace it with reset
In practice over the years I never really ended up running init by
itself because I always wanted my database seeded too.

Now it'll be standard to just run a flask db reset to reset everything.

If in the future a use case pops up where dropping, creating and seeding
the database needs to be separated out it will be easy to add new drop
and create commands, and then they can be called as needed.