Add init command that is similar to alembic init
Since alembic has an init command to generate a bunch of config files it
felt natural to use init here too. You would use this as a drop in
replacement for alembic init.

In addition to generating alembic configs this also creates an initial
seeds.py file to use as an example for seeding your database.

Most of the alembic config files are similar to the defaults. The main
changes are for dynamically reading the SQLAlchemy URI and supporting a
custom location for where the alembic directory is.