Better document revision --autogenerate
Alembic's list of what autogenerate can and cannot detect is changing
all the time.

In the past indexes were not detected, but that's no longer true. They
were added quite some time ago.

I've gotten burned by this in the past and kind of wrote off ever using
autogenerate but as of 1.4.x the list of deteted changes is really
solid.

I don't think they should be enabled by default at this tool's level
since there's very big implications of using this feature without
knowing the limitations but perhaps in the future a config option could
be added to control this behavior.